+++ src/image.rs
@@ -11,8 +11,46 @@ use std::os::raw::c_int;
 
 const MAX_IMAGE_SIZE: u32 = std::i32::MAX as _;
 
+pub struct Plane<'a> {
+    pub data: &'a [u8],
+    pub width: u32,
+    pub height: u32,
+    pub stride: usize,
+    pub bits_pre_pixel: u8,
+}
+
+pub struct PlaneMut<'a> {
+    pub data: &'a mut [u8],
+    pub width: u32,
+    pub height: u32,
+    pub stride: usize,
+    pub bits_pre_pixel: u8,
+}
+
 pub struct Image {
     pub(crate) inner: *mut heif_image,
+}
+
+pub struct Planes<'a> {
+    pub y: Option<Plane<'a>>,
+    pub cb: Option<Plane<'a>>,
+    pub cr: Option<Plane<'a>>,
+    pub r: Option<Plane<'a>>,
+    pub g: Option<Plane<'a>>,
+    pub b: Option<Plane<'a>>,
+    pub a: Option<Plane<'a>>,
+    pub interleaved: Option<Plane<'a>>,
+}
+
+pub struct PlanesMut<'a> {
+    pub y: Option<PlaneMut<'a>>,
+    pub cb: Option<PlaneMut<'a>>,
+    pub cr: Option<PlaneMut<'a>>,
+    pub r: Option<PlaneMut<'a>>,
+    pub g: Option<PlaneMut<'a>>,
+    pub b: Option<PlaneMut<'a>>,
+    pub a: Option<PlaneMut<'a>>,
+    pub interleaved: Option<PlaneMut<'a>>,
 }
 
 pub struct ScalingOptions {}
@@ -84,6 +122,74 @@ impl Image {
         })
     }
 
+    fn plane(&self, channel: Channel) -> Option<Plane> {
+        if !self.has_channel(channel) {
+            return None;
+        }
+
+        let width = self.width(channel).unwrap();
+        let height = self.height(channel).unwrap();
+        let bits_pre_pixel = self.bits_per_pixel(channel).unwrap();
+        let mut stride: i32 = 1;
+        let data = unsafe { heif_image_get_plane(self.inner, channel as _, &mut stride) };
+        let size = height as usize * stride as usize;
+        let bytes = unsafe { slice::from_raw_parts(data, size) };
+        Some(Plane {
+            data: bytes,
+            bits_pre_pixel,
+            width,
+            height,
+            stride: stride as _,
+        })
+    }
+
+    fn plane_mut(&self, channel: Channel) -> Option<PlaneMut> {
+        if !self.has_channel(channel) {
+            return None;
+        }
+
+        let width = self.width(channel).unwrap();
+        let height = self.height(channel).unwrap();
+        let bits_pre_pixel = self.bits_per_pixel(channel).unwrap();
+        let mut stride: i32 = 1;
+        let data = unsafe { heif_image_get_plane(self.inner, channel as _, &mut stride) };
+        let size = height as usize * stride as usize;
+        let bytes = unsafe { slice::from_raw_parts_mut(data, size) };
+        Some(PlaneMut {
+            data: bytes,
+            bits_pre_pixel,
+            width,
+            height,
+            stride: stride as _,
+        })
+    }
+
+    pub fn planes(&self) -> Planes {
+        Planes {
+            y: self.plane(Channel::Y),
+            cb: self.plane(Channel::Cb),
+            cr: self.plane(Channel::Cr),
+            r: self.plane(Channel::R),
+            g: self.plane(Channel::G),
+            b: self.plane(Channel::B),
+            a: self.plane(Channel::Alpha),
+            interleaved: self.plane(Channel::Interleaved),
+        }
+    }
+
+    pub fn planes_mut(&mut self) -> PlanesMut {
+        PlanesMut {
+            y: self.plane_mut(Channel::Y),
+            cb: self.plane_mut(Channel::Cb),
+            cr: self.plane_mut(Channel::Cr),
+            r: self.plane_mut(Channel::R),
+            g: self.plane_mut(Channel::G),
+            b: self.plane_mut(Channel::B),
+            a: self.plane_mut(Channel::Alpha),
+            interleaved: self.plane_mut(Channel::Interleaved),
+        }
+    }
+
     pub fn has_channel(&self, channel: Channel) -> bool {
         unsafe { heif_image_has_channel(self.inner, channel as _) != 0 }
     }
@@ -111,7 +217,7 @@ impl Image {
         Ok(Image { inner: image })
     }
 
-    pub fn add_plane(
+    pub fn create_plane(
         &mut self,
         channel: Channel,
         width: u32,
@@ -128,24 +234,6 @@ impl Image {
             )
         };
         HeifError::from_heif_error(err)
-    }
-
-    pub fn plane_mut(&mut self, channel: Channel) -> Result<(&mut [u8], usize), HeifError> {
-        let height = self.height(channel)? as usize;
-        let mut stride: i32 = 1;
-        let data = unsafe { heif_image_get_plane(self.inner, channel as _, &mut stride) };
-        let size = height * (stride as usize);
-        let bytes = unsafe { slice::from_raw_parts_mut(data, size) };
-        Ok((bytes, stride as usize))
-    }
-
-    pub fn plane(&self, channel: Channel) -> Result<(&[u8], usize), HeifError> {
-        let height = self.height(channel)? as usize;
-        let mut stride: i32 = 1;
-        let data = unsafe { heif_image_get_plane_readonly(self.inner, channel as _, &mut stride) };
-        let size = height * (stride as usize);
-        let bytes = unsafe { slice::from_raw_parts(data, size) };
-        Ok((bytes, stride as usize))
     }
 
     //    TODO: need implement