+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added small example of usage into README.md.
 - Changed some enum values and name of methods to comply with the Rust
   naming conventions.
+- Methods ``Encoder::set_lossless()`` and ``Encoder::set_lossy_quality()``
+  replaced by ``Encoder::set_quality()``.
+- Added methods ``Image::planes()`` and ``Image::planes_mut()``.
 
 ## [0.1.0]